requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.11
django-cors-headers==3.10.0
django-filter==2.4.0
django-nested-admin==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.8.5
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.1.0
